requirements.txt
langchain 
wikipedia 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
langchain-community 

 